Write time as UnixTime to the db, fixing serialization issues

+++ AirCombatMatchmakerBot/Data/Leagues/LeagueData/LeagueDataComponents/MatchScheduler.cs
@@ -27,7 +27,7 @@ public class MatchScheduler : logClass<MatchScheduler>
     public MatchScheduler() { }
 
     // TODO: Implement this method so it's executable from some command that admin can use (for initiation of a season, for example)
-    public void ActivateMatchScheduler(int _duration)
+    public void ActivateMatchScheduler(ulong _duration)
     {
         if (MatchSchedulerActive)
         {
